+++ routes/community.router.js
@@ -4,8 +4,7 @@ import * as articleReadController from "../controllers/community/article.read.co
 import * as commentController from "../controllers/community/comment.community.controller.js";
 import * as articleLikeController from "../controllers/community/article.like.community.controller.js";
 import * as commentLikeController from "../controllers/community/comment.like.community.controller.js";
-import * as articleReportController from "../controllers/community/article.report.community.controller.js";
-import * as commentReportController from "../controllers/community/comment.report.community.controller.js";
+import * as reportController from "../controllers/community/report.community.controller.js";
 import * as articleAggregateController from "../controllers/community/article.aggregate.community.controller.js";
 // 사용자 인증 미들웨어 (추후 네임스페이스 방식으로 변경 필요)
 import * as articleValidator from "../utils/validators/community/article.validators.js";
@@ -20,6 +19,12 @@ const router = Router();
 /*
   관리자 메뉴 : 게시판 생성
 */
+
+router.post(
+  "/",
+  authenticateAccessToken,
+  articleCrudController.createCommunity
+);
 
 /*
   게시판 조회
@@ -162,14 +167,14 @@ router.delete(
 router.post(
   "/:communityId/articles/:articleId/report",
   authenticateAccessToken,
-  articleReportController.reportArticle
+  reportController.reportArticle
 );
 
 // 댓글을 신고합니다.
 router.post(
   "/:communityId/articles/:articleId/comments/:commentId/report",
   authenticateAccessToken,
-  commentReportController.reportComment
+  reportController.reportComment
 );
 
 /*
